Move reusable translation-related entity methods to Translatable trait
