+++ src/Pim/Bundle/CatalogBundle/Entity/Family.php
@@ -69,16 +69,6 @@ class Family implements TranslatableInterface, ReferableInterface, VersionableIn
     }
 
     /**
-     * Returns the label of the family
-     *
-     * @return string
-     */
-    public function __toString()
-    {
-        return $this->getLabel();
-    }
-
-    /**
      * Get id
      *
      * @return integer
@@ -266,32 +256,6 @@ class Family implements TranslatableInterface, ReferableInterface, VersionableIn
                 );
             }
         )->toArray();
-    }
-
-    /**
-     * Get label
-     *
-     * @return string
-     */
-    public function getLabel()
-    {
-        $translated = $this->getTranslation() ? $this->getTranslation()->getLabel() : null;
-
-        return ($translated !== '' && $translated !== null) ? $translated : '['.$this->getCode().']';
-    }
-
-    /**
-     * Set label
-     *
-     * @param string $label
-     *
-     * @return Family
-     */
-    public function setLabel($label)
-    {
-        $this->getTranslation()->setLabel($label);
-
-        return $this;
     }
 
     /**
