+++ src/Pim/Bundle/CatalogBundle/Model/AbstractAttribute.php
@@ -637,16 +637,6 @@ abstract class AbstractAttribute implements TimestampableInterface, Translatable
     }
 
     /**
-     * To string
-     *
-     * @return string
-     */
-    public function __toString()
-    {
-        return $this->getLabel();
-    }
-
-    /**
      * Get group
      *
      * @return AttributeGroup
@@ -1236,32 +1226,6 @@ abstract class AbstractAttribute implements TimestampableInterface, Translatable
     public function getTranslationFQCN()
     {
         return 'Pim\Bundle\CatalogBundle\Entity\AttributeTranslation';
-    }
-
-    /**
-     * Get label
-     *
-     * @return string
-     */
-    public function getLabel()
-    {
-        $translated = ($this->getTranslation()) ? $this->getTranslation()->getLabel() : null;
-
-        return ($translated !== '' && $translated !== null) ? $translated : '['.$this->getCode().']';
-    }
-
-    /**
-     * Set label
-     *
-     * @param string $label
-     *
-     * @return string
-     */
-    public function setLabel($label)
-    {
-        $this->getTranslation()->setLabel($label);
-
-        return $this;
     }
 
     /**
